requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.1
bcrypt==3.2.0
blessed==1.18.1
boto3==1.18.18
botocore==1.21.18
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
cfn-flip==1.2.3
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
cryptography==3.4.7
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
durationpy==0.5
future==0.16.0
gunicorn==20.1.0
hjson==3.0.2
idna==3.2
jmespath==0.10.0
jsonschema==3.2.0
kappa==0.6.0
numpy==1.21.1
opencv-python-headless==4.5.3.56
paramiko==2.7.2
pathspec==0.5.9
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
tomli==1.2.1
tqdm==4.62.0
troposphere==3.0.2
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.53.0
